Add files via upload
Implement an event driven program that allows user to select rock, paper or scissors and a computer opponent that randomly generates rock, paper or scissors when user plays. In addition, implement an extra feature. In this case, I implemented a score board that tracks the score. I used HTML, CSS and Javascript to create this program.